refactor: dto 구조 개선
