--- doonut-app-external-api/src/main/kotlin/com/doonutmate/fcm/client/dto/Message.kt
@@ -1,6 +1,0 @@
-package com.doonutmate.fcm.client.dto
-
-data class Message(
-    val notification: Notification,
-    val token: String,
-)
