--- doonut-app-external-api/src/main/kotlin/com/doonutmate/fcm/client/dto/Notification.kt
@@ -1,6 +1,0 @@
-package com.doonutmate.fcm.client.dto
-
-data class Notification(
-    var title: String,
-    val body: String,
-)
